refactor: the architecture of the project
