+++ src/app/app.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common';
-import { AppService } from 'src/services/app.service';
+import { AppService } from 'src/app/app.service';
 
 @Controller()
 export class AppController {
